+++ matfealib/plot/__init__.py
@@ -1,0 +1,2 @@
+"""Sup-package for visualizing atomic features at a periodic table."""
+